added Course.viewblob
The basic idea is to return a read-only dictionary for read-only access.
However since access is slow, the dictionary shall be lazy. The lazyness
happens inside view.LazyView. So now there is Course.viewblob which
returns a dict with the same keys as the previous Blob attributes. Since
everything is lazy, we can turn all getblob and getmetablob calls into
viewcalls when teaching all call sites to look at entries and not
attributes. Thankfully jinja treats them almost the same.