Have a set-like semantics for the attached blobs
That is, relinking the same blob to the same page is idempotent. Markus
convinced me that this is waht the users expect -- by prommsing that he
himself will personally take care of all mails by users complaining that
it does not behave like a list, where adding the same object is not idempotent.