parser: use an output stack
Our output previously was a sequence of characters. Now it is a stack of
these outputs corresponding to the state stack. Each time we pop a
stack, we call handle_statename on the output stack top item. This
permits abstracting part of the output generation from the parsing. In
addition this technique somehow speeds up the parser by a factor of
almost 2.