It would be nice, if downloaded images would be advertised as images.
This has the problem, that we need to guess the file type of a blob, but I
think this would be worth the effort, since it makes the handling easier in
most cases (i.e. viewing the image after downloading, since most browsers
offer to open an image viewer if you download an image).