added some experimental estimator
We estimate a document by parsing it and assigning different weights to
different structures. So all we have to do is add yet another BaseParser
subclass and write some handle_* methods, right? Wait no. The BaseParser
assumes strings to be returned and will add further strings to the
results. Damn. But, does it actually do anything else with those
strings? No? So everything that provides __add__ looks like a string?
Great. Let's duck-type an Estimate class that pretends to be string, but
actually wraps an int.