do_attachblob: improve exception handling
Previously we would assume that if an exception was raised from
attachblob it was "probably" due to an invalid filename (and even before
we checked the exception message). What we really want is to know what
failed. This is what exception hierarchies are for. So we add an
InvalidBlobFilename exception inheriting from CheckError. Now we can
except this exception and act accordingly.