Validate rcs versions passed as input to endedit
Validate user-provided rcs version strings. Not that rcs wouldn't
do that itself, but in this way, we never pass invalid strings to
a subprocess -- as this might cause some system administrators to
panic (though not being a security issue). In that way we also never
pass very long strings to a subprocess.check_call and therefore don't
have to rely on this being implemented properly either.