replace the NotFound, inherited from checkexsitence branch, by Conflict
...as a better approximation of what we want: a transient (hence
non-cachable) error. The description says "The request could not be
completed due to a conflict with the current state of the resource."
Admittedly, it is more intended for conflicting edits (where we give a
positive answer at http-level), but in a sense we more or less should
care only about non-cachability (so that honest users don't suffer
collateral dammage); a malicious user will make his decission on whether
to resubmit independently of what we tell him to do anyway.

Actually, 500 might be an option as well ;-)