sqlite3: SELECTs also need to be committed
This fix should solve the (compacted) traceback applying to
1e1b196edd6aa52f6ac5fd6d49ccc2426106fc54 below.

Traceback (most recent call last):
  ...
  File ".../dokuforge/application.py", line 75, in get
    self.expire(now)
  File ".../dokuforge/application.py", line 135, in expire
    (now - self.expire_after,))
OperationalError: database is locked

To reproduce present the database with a non-existent cookie. This will cause a
SELECT statement to be issued (from get) without it being followed by an UPDATE
(and thus commit). Then cause expire to be called for the same non-existent
cookie by sending the request again. (Hope I didn't get this wrong.)