Simplify interface: remove parameter initialline
the (optional, named) parameter "initialline" was used to have a
shorter notation for

x = linegroup()
liengroup.appendline(line)

However, as it turns out, this shorthand notation made the
interface much harder to understand. So get rid of it and
expand the abbreviation at the two (2) places where it
was used.