fix testsuite with werkzeug 0.8
The werkzeug 0.8 release starts with correctly quoting URLs it emits.
This results in ! being quoted as %21. Back in the old CGI standard
(which wsgi is based upon) the server would do the unquoting and the
PATH_INFO variable would contain the unquoted version! So our stub
server needs to do this unquoting as well. In Addition our url matches
need to be adapted to match urls with %12. Thanks to Armin Ronacher for
explaining this subtlety.