added class TarWriter
The TarWriter class collects the tar*Chunk functions as methods and
preserves state between two operations. That way it lifts the assumption
on tarfile being unbuffered. A side effect of this change is that it is
now trivial to add gzip or bzip2 compression. Just change the mode to
"w|gz" or "w|bz2".