Add timestamp functions to Academy and Course
This will allow us to replace the commit info for courses and academies with
a more leigthweight alternative.

This has certain tradeoffs since this relies on the mtime of some files,
which is not guaranteed to be entirely accurate, but it is accurate in the
sense of "if the mtime didn't change, then nothing changed".