createexample: move the long test text to size>0
The reason for adding sizes was to reduce the time to create the
examples. Now the time is largely dependent on the number of rcs calls.
The size of a page is mostly irrelevant. Given that the example pushes
test coverage by some 100 lines it seems reasonable to have a slightly
larger "small" example.