move testparser.py inside test.py
This way we can include its findings in the test report and the coverage
report. While moving I also discovered that some characters were
missing, so I added them. And voila:

AssertionError: '\n\n$&&>a--]( .a_<((]]\n_1{\\ a\\&1&a> >-_* -}-a[1}$<&_"]&-&a-A]{>.)\n1\n[[_$" ){(1\\<{<<<$\n' != '\n\n$&&>a--]( .a_<((]]\n_1{\\ a\\&1&a> >-_* -}-a[1}$<&_"]&-&a-A]{>.)\n_\n\n\n[[_$" ){(1\\<{<<<$]]'