Add estimate functionality to the Outline class.
This will allow to display estimates on course overview pages. This is a
highly desired feature to make the work of the Dokuteam easier.

This way (i.e. stuffing it into Outline) the ressource consuming parsing is
done just once for determination of
a) Headings,
b) Estimates.

This obsoletes dfOverview which was a small performance enhancement.