Make forbidden not insert a superfluous pair of braces.
The additional pair of braces means lots of work for editors (since it has
to be removed) and interferes with the '\forbidden' macro.

I think the idea was, to shield the thing which was being forbidden from
outside influence. Since TeX command sequences are composed only of letters
the EscapeCommands feature is guaranteed to catch them all and no sideeffect
is possible, hence the pair of braces can go.