Add more commands to the whitelist.
This adds the command list from the first installment of the exporter in the
master-2012-07 branch. This should cover 95% of the use cases.

The mathbb macro is removed from the allowed list, since it is not a
semantic macro (yeah I see the use case where it actually becomes a semantic
macro, but I would prefer to err on the side of caution and from experience
I would extrapolate this not to be too cumbersome), but influences the
formatting of elements.