converted action links to fake forms
Added macro fakebutton to base.html and converted some of the links to
use fakebutton. Those that seemed useful. Most of the time the following
sed expression helped.

s#<a href="{{ *\(.*\)|e *}}">\(.*\)</a>#{{ fakebutton(\1, "\2") }}#

The devil is in the detail though. (Malformed html, &ouml;, etc.)

Also note that edit.html looks broken due to crazy absolute positioning
in the css. I failed to fix this.