HTML/4.0 compliance
The most common error was wrong nesting of text elements. So buttons are
considered text and must be in a some div, p, h1, or something else.
Similarly text. On the other hand a p cannot contain a div. There are
still problems. For instance the script tag for the mathjax cookie fails
to validate, but I was unable to fix it. Not all pages or features were
tested. I probably broke some CSS.