refactor the microtype feature interface
Instead of a class with static methods we now go for a function taking a
str and returning an iterable of str. Then the application of features
becomes what is known as "foldl (>>=)" in Haskell world. Things
generally become shorter. Just doMicrotype was eliminated as a special
case as well. It just applies the additional SplitSeparators feature.

Thanks to Klaus for explaining what I am doing here.