Implement more tests for exporter microtypography.
This should cover most of the desired behaviour of the microtypography if
not all of it. I expect all of the tests to break right now, because there
are some behaviour changes (like \@ added). Some of the behaviour may be
omitted in the near future due to implementation cost; these should then be
marked as 'expected failure' and supplemented with a new test specifying the
current behaviour.