Shorten Stack-Trace.
Just store the last three lines of stack trace. The whole stacktrace is
really ugly if piped to syslog, since syslog squelches it onto a single
line. Thus we keep just the last three lines which should contain the exact
location of the error (file and line) -- which is the most interesting
information besides the error message. Hence we have a better SNR.