Make mathbutton into link.
Just as with fakebutton. This basically unbreaks the test suite. The
'text-decorate: none' seems to be necessary for Firefox 10 but not for
Firefox 15 -- weird, but everybody says, that consisten behaviour of
browsers is non-existent and it seems to be correct ..

There are still lots of failing tests since many links were converted into
buttons and thus the test suite doesn't find them.