convert str to bytes and unicode
Since Py2.6 bytes is an alias for str and string literals can also be
written as b"foo". In Py3.x bytes and str will have a different meaning.
Also the name unicode will be gone and called str instead, but the 2to3
tool is capable of handling this.

So wherever we want to talk about sequences of octets, we should use
bytes and b"literals". And where ever we mean unicode strings we should
use unicode and u"foo" (which will be transformed to str and "foo" by
2to3).

One notable difference is that bytes (as for Py3.x) does not support
formatting via __mod__. So this needs to be replaced. Either use + or
use a unicode string and encode it.

This commit is incomplete. I converted a number of places, but as can be
seen by the number of literals lacking a u or b prefix and the number of
uses of str, there is still a lot to be done.

For testing a single file part of the Py3.x behaviour can be reproduced
by adding the following line to the top of a module:

from __future__ import unicode_literals; str = unicode