Provide a "course number" for each course in the export
Our macros for typestting require \course to have a course number
as its first argument. The course data structure, however, does not
have such a number. In fact, it is a feature that courses can be accessed
by a proper name, rather than just a number. To make up for this, we
guess the number from the name if possible (e.g., in storagedir names
like course16) and take 0 otherwise.