Add interstitial page before deleting a logical page
We're not really afraid of users accidentally deleting pages they
still need; in fact, every delete can easily be undone by the
user. The main reason for this interstial page is to have the same
control flow as edit -- which is an indirect action by nature (you
have to provide additional information, the new version of the text).