discard FileDoesNotExist
It was thrown at only one place, converted to StorageFailure in the only
caller and sometimes raised when the file actually existed. rlogv now
raises StorageFailure instead of returning None. Thus rloghead does not
need to convert that None into FileDoesNotExist anymore and commitstatus
does not have to eliminate that FileDoesNotExist anymore. The status
method was adapted to the new rlogv, but it could use some API change as
well.