test suite is broken when http_proxy is set
urllib queries that environment variable to detect proxies. urllib2 then
installs a ProxyHandler that mechanize uses. Even though we never query
the proxy, the PATH_INFO variable is then set to
"http://www.dokuforge.de/something" instead of just "/something". This
is caught by wsgiref.validate to make all mechanize tests fail. The
solution is to not use a proxy in any case.