Restrict in-course navigation to non-deleted pages
It is a feature of dokuforge that (currently) deleted pages can be
viewed and edited normally (this also avoids unnecessary data loss for
edit-delete conflicts). Of course, for those pages, there is no
well-defined "previous" or "next" page, so only show the in-course
navigation for non-deleted pages. This also avoids an exception thrown
by course.pages.index(page).