fix accidental integer division
The standard division (a / b) changes for py3k. Dividing two integers
would previously result in an integer and will then result in a float.
Now in our method fullline we would accidentally use integer division:

    math.ceil(self.weightedchars / self.charsperline)

To fix this once and for all, this commit changes the dividends to
always be float objects avoiding the need for int division (and the
nasty DeprecationWarning) altogether.

After this commit, no direct code from df2 causes a DeprecationWarning
when run with python -3 test.py.