Only consider 0-9 as the forming parts of numerals
In the microtypography for numerals, we consider a
numeral what is formed of digits; numerals are internally
converted to int before formatted with correct kerning.
However, there are characters, most notably superscript-3
(i.e., u'\xb3') that are digits in the sense of isdigit(),
but cannot be parsed as int. So restrict the special treatment
for numerals to strings formed only of 0-9.