work around webtest 1.3.4 limitation
The mentioned version treats <select ... multiple> as a non-multiple
Select, which causes later failures in the test suite. This is fixed in
webtest as https://github.com/Pylons/webtest/pull/62. Thanks to Bjoern
Bastian for the report and testing. This commit can be reverted once we
stop supporting this version of webtest, which probably amounts to the
point of no longer supporting Debian wheezy.