fix DokuforgeWebTests.testStyleguide for py3
We pass a unicode topic to Application.do_styleguidetopic. Then we turn
it into a path. Unfortunately paths use str (i.e. a Python version
dependent type). Thus the encoded bytes topic is not present in the
directory listing. Rather we need to prohibit the encoding operation for
Python3. This also fixes our invocation of
Application.render_styleguide. It is documented to take a unicode topic,
but we were passing bytes all the time.