storage.py: in call to ci convert user name
The arguments to subprocess are, for whatever reason, unicode
strings. So, while we have the user name as bytes, which
we want to eventually be passed to execve(2), we still have
to convert it to unicode first. Fortunately, our user names
are 7-bit clean anyway, so that decoding (and the reencoding by
subprocess) shouldn't do too much harm.

Signed-off-by: Klaus Aehlig <aehlig@linta.de>