change storage.Storage.store parameter message to unicode
No caller actually provides the message, so no code needs to be updated
elsewhere. The message is encoded to bytes using utf8 before passing it
to ci. While not documented, the subprocess module deals correctly with
byte. Any str arguments are converted to bytes using the filesystem
encoding, but we don't want the filesystem encoding here.