fix representation of "no lines"
The b"\n".join(lines) + b"\n" technique was mis-representing "no lines"
as a single b"\n". When parsing the Index file on Python 3, it would
come up with a single empty line and turning that into an int would
fail.

When serializing no lines, we should return empty bytes, not b"\n".