course.html: use loop.index instead of counting by hand
So far, we computed the loop index by hand in the template,
using a construction as follows

  {%- set i = 0 -%}
  {%- for page in course.outlines %}
    {%- set i = i + 1 -%}

    ... {{ i | e }}
  {%- endfor -%}

However, it seems that the scoping rules for i have changed with
the new version of the template engine and are now 0 at the beginning
of every round in the loop. Fortunately, we don't need to count by
hand and can simply use loop.index.