application: use keyword-only arguments
We use the following pattern a lot:

    def somefunction(arg, kwarg1=None, kwarg2=None, ...):
        assert kwarg1 is not None and kwarg2 is not None and ...
        ...

The intention was to have keyword-only arguments, but Python 2 didn't
allow us to do so. Now after dropping Python 2.X support, we can
actually use keyword-only arguments as a language feature.

The old way is also problematic for type checking with mypy. Since we
are assigning a default of None, the type must allow None. So we have to
wrap all the types in typing.Optional. Dropping the None default removes
the need for this wrapper.