avoid changing types of variables
mypy is notoriously bad at understanding code where a variable contains
objects of multiple times. A common workaround is renaming one of the
involved variables. This commit does so in a number of places. In other
occasions, the need for a variable is eliminated. In all cases, an
assignment to an existing variable with a value of different type was
removed.