fix str % bytes interpolation in Storage.endedit
Now that we know that Storaget.status returns bytes, currentversion also
becomes bytes. The user variable is declared optional bytes in the
method parameters. fullpath also returns bytes. So args is a list of
mixed str and bytes and we interpolate some str with bytes values. As in
other commands, we should use bytes exclusively and avoid the
unintentional quotes and b prefix around our interpolations.