require itemize and enumerate to be exported with "[joinedup,packed]"
The most common types of itemized and enumerated lists are without options
and with these options. As it is easier to see where they are not needed
and remove them than to look up what the options should be and add them,
it would make sense to add them by default.