base: refactor base template by using bootstraps grid system
This does not attempt to do absolute placements of header and footer,
but uses bootstraps containers instead. So, they do not stay on the page
and cover the actual content.

The header is now slimmer, containing only the title of the page,
the navigation, the user's name, logout button and references to
the style sheet and the formula preview link.

The links to admin and group page are now on the index page.

The footer is now centered.