ci: add basic ci setup
This adds a requirements.txt file containing the versions of the
python packages. The current versions are temporary and needs to be
adjusted to reflect the true version numbers.

The version of python we use for testing (currently python3.9) is also
temporaly and needs to be adjusted once we agreed upon a minimal python
version.

closes #5