Link URLs in web view
This uses the urlize jinja filter to link the urls in web view.
Doing so ourselves in the parser would have been unnecessarily
complex, and have required additional escaping of quotation marks
etc.

While user content is unsafe, the worst this can lead to is broken
links. For some reason, even though autoescaping is disabled,
urlize does not seem to be aware of this, so the content needs to
be marked as safe explicitly (which does not change anything
otherwise)

However, since this leads to a different url detection algorithm
for export and web, this allows for edge cases where these do not
match.